almost all commands, still borked
