+++ src/classes/handlers/index.ts
@@ -1,6 +1,3 @@
-export * from './CommandHandler.js';
-export * from './ComponentHandler.js';
 export * from './CooldownHandler.js';
 export * from './EventHandler.js';
-export * from './ModalHandler.js';
 export * from './InteractionHandler.js';
