+++ src/commands/case/view.ts
@@ -1,7 +1,7 @@
 import { Embed, type FluorineClient } from '#classes';
 import { type ChatInputCommandInteraction, SlashCommandSubcommandBuilder } from 'discord.js';
 
-export async function run(client: FluorineClient, interaction: ChatInputCommandInteraction) {
+export async function onSlashCommand(client: FluorineClient, interaction: ChatInputCommandInteraction) {
     const id = interaction.options.getInteger('id');
     const caseObj = await client.cases.getOne(interaction.guildId, id);
 
@@ -33,7 +33,7 @@ export async function run(client: FluorineClient, interaction: ChatInputCommandI
     interaction.reply({ embeds: [embed] });
 }
 
-export const data = new SlashCommandSubcommandBuilder()
+export const slashCommandData = new SlashCommandSubcommandBuilder()
     .setName('view')
     .setNameLocalizations({ pl: 'zobacz' })
     .setDescription('Check a moderation case')
