+++ src/commands/dev/sql.ts
@@ -17,7 +17,7 @@ import type { NonCommandInteractionData } from '#types';
 export async function onSlashCommand(client: FluorineClient, interaction: ChatInputCommandInteraction) {
     const modal = new ModalBuilder()
         .setTitle('Evaluate')
-        .setCustomId(`nsql:${interaction.user.id}`)
+        .setCustomId(`sql:${interaction.user.id}`)
         .addComponents(
             new ActionRowBuilder<TextInputBuilder>().addComponents(
                 new TextInputBuilder()
@@ -58,7 +58,7 @@ export async function onModal(
 
 export const modalData: NonCommandInteractionData = {
     exists: true,
-    name: 'nsql'
+    name: 'sql'
 };
 
 export const slashCommandData = new SlashCommandSubcommandBuilder()
