+++ src/commands/warn.ts
@@ -37,6 +37,8 @@ export async function onSlashCommand(client: FluorineClient, interaction: ChatIn
             { name: 'CASE_ID', value: caseObj.caseId.toString() }
         ]);
 
+    console.log('hi');
+
     interaction.reply({ embeds: [embed] });
     client.cases.logToModerationChannel(interaction.guildId, caseObj);
 }
