+++ src/util/index.ts
@@ -1,7 +1,5 @@
-export * from './avatar.js';
 export * from './clean.js';
 export * from './command.js';
 export * from './files.js';
 export * from './fragmentText.js';
-export * from './info.js';
 export * from './splitArray.js';
